// src/api/content/types/page.ts
export interface CmsPage {
  content: PageContent;
  pageHeader?: StringValue;
  headerIcon?: FileValue;
  pageUrl: StringValue;
  pageKey: StringValue;
  journeyStage?: NumberValue;
  journeyType?: JourneyType;
  journeyStepNumber?: NumberValue;
  journeyHideBackButton?: BooleanValue;
  showAsStickOut?: BooleanValue;
  backPageKey?: StringValue;
  heroBlocks?: { values: HeroBlock[] };
  pageMenu?: { value: PageContentValues };
  removeFromJourneySteps?: BooleanValue;
  showDialogOnLoad?: DialogElement;
  hideTemplateNavigation?: BooleanValue;
  pageBackground?: BackgroundColorConfig;
}

export interface HeroBlock {
  elements: {
    heroContent: {
      value: PageContentValues;
    };
    heroImage?: FileValue;
  };
  type: string;
}

export type ThemeBackgroundColorSelection = SelectionValue<AppColorType>;
// ...
export type JourneyTypeSelection =
  | 'retirement'
  | 'quoteselection'
  | 'bereavement'
  | 'transfer2'
  | 'Identity'
  | 'dcretirementapplication'
  | 'dbcoreretirementapplication';

export type AppColorType =
  | 'Primary'
  | 'Secondary'
  | 'SecondaryDark'
  | 'Tertiary'
  | 'TertiaryDark'
  | 'Support60'
  | 'Support60Dark'
  | 'Support80'
  | 'Support80Dark'
  | 'White'
  | 'Black'
  | 'None';

export type PanelLayoutSelection = SelectionValue<
  '50/50' | '80/20' | '20/80' | '45/10/45' | '66/33' | '33/33/33' | '100'
>;

type JourneyType = SelectionValue<JourneyTypeSelection>;

export interface PageContent {
  values?: PageContentValues[] | null;
}
export interface PageContentValues {
  elements: PageElements;
  type: string;
  name: string;
}

export interface PanelListItem {
  elements: {
    columns: { values?: PanelColumn[] };
    header?: StringValue;
    layout?: PanelLayoutSelection;
    reverseStacking?: BooleanValue;
    panelKey?: StringValue;
  };
}

interface PageElements
  extends Partial<ButtonElements>,
    Partial<ResourceListElements>,
    Partial<ResourceListItemElements>,
    Partial<InfoTileElements>,
    Partial<ChartElements>,
    Partial<DataSummaryElements>,
    Partial<TimelineElements>,
    Partial<DataTableElements>,
    Partial<TrackerElements> {
  textLabel?: StringValue | null;
  type?: SelectionValue;
  themeColorForBackround?: ThemeBackgroundColorSelection | null;
  themeColorForBackground?: ThemeBackgroundColorSelection | null;
  elementColor?: ThemeBackgroundColorSelection | null;
  elementsColor?: ThemeBackgroundColorSelection | null;
  backgroundColor?: ThemeBackgroundColorSelection | null;
  gridBackgroundColour?: StringValue;
  blockHeader?: BlockHeader | null;
  content?: StringValue | null;
  header?: StringValue;
  panelNameLabel?: StringValue;
  formKey?: StringValue;
  panelList?: { values?: PanelListItem[] };
  panels?: { values?: PanelListItem[] };
  labels?: { values?: LabelValues[] };
  tooltips?: { values?: TooltipValues[] };
  buttons?: { values?: ButtonValues[] };
  linkGroups?: { values?: LinkGroup[] };
  cards?: { values?: CardsValues[] };
  text?: StringValue;
  reverseStacking?: BooleanValue;
  layout?: PanelLayoutSelection;
  panelKey?: StringValue;
  columns?: { values?: PanelColumn[]; value?: number };
  rows?: { value?: number };
  answers?: Answers;
  questionKey?: StringValue;
  questionText?: StringValue;
  bottomPanel?: { value?: PanelListItem };
  showInDropdown?: BooleanValue;
  hideSaveAndExit?: BooleanValue;
  disableCallToAction?: BooleanValue;
  contentBlockKey?: StringValue;
  openFile?: FileValue;
  journeyPage?: JourneyPage;
  stage?: Parameters;
  maxNumberOfJourneySteps?: StringValue;
  pageKey?: StringValue;
  customActionKey?: StringValue;
  analyticsKey?: StringValue;
  headerLink?: StringValue;
  subHeader?: StringValue;
  parameters?: Parameters;
  showAlwaysOnTop?: BooleanValue;
  callToAction?: CallToAction;
  callToActionIfData?: CallToAction;
  backgroundColour?: StringValue;
  showInAccordion?: SelectionValue;
  linkText?: StringValue;
  pageMenuItem?: Parameters;
  orderedListItems?: { values?: { elements: OrderedListItemElement }[] };
  orderedListKey?: StringValue;
  hideNumber?: BooleanValue;
  defaultItemImage?: FileValue;
  showAllItems?: BooleanValue;
  items?: StringValue;
  avoidBranching?: BooleanValue;
  actionButtons?: CallToAction;
  checkbox?: { values: Checkbox[] };
  checkboxListKey?: StringValue;
  description?: StringValue;
  mainContent?: StringValue;
  chart?: MixedValue<{ elements: ChartElements } | undefined>;
  footerIcon?: FileValue;
  caption?: StringValue;
  captionIfData?: StringValue;
  taskIconIfDone?: StringValue;
  captionIfZero?: StringValue;
  iconIfZero?: StringValue;
  taskIconIfToDo?: StringValue;
  taskIconIfZero?: StringValue;
  taskLimit?: NumberValue;
  tasks?: TasksValue;
  gridKey?: StringValue;
  contentBlocks?: { values: PageContentValues[] };
  panel?: { value: PanelListItem };
  errorContent?: StringValue;
  dataSourceUrl?: StringValue;
  errorText?: StringValue;
  alternateTableStyle?: SelectionValue<'Transparent'>;
  badgeKey?: StringValue;
  addBorder?: BooleanValue;
  showInAlternateStyle?: BooleanValue;
  hideCloseInAlternateStyle?: BooleanValue;
}
// ...
export type PanelSideSelection = 'right' | 'left' | 'top' | 'bottom';

export interface PanelColumn {
  alignment: { elementType: 'optionselection'; value: ColumnAlignment };
  themeColorForBackround?: ThemeBackgroundColorSelection | null;
  contentBlocks?: ContentBlocksListItems;
  enablePadding?: BooleanValue;
  roundCorners?: BooleanValue;
  reducedGap?: BooleanValue;
  border?: MultiSelectionValue<PanelSideSelection>;
  paddingSide?: MultiSelectionValue<PanelSideSelection>;
  desktopPaddingMultiplier?: NumberValue;
}

export interface ColumnAlignment {
  label: string;
  selection:
    | 'top-left'
    | 'top-center'
    | 'top-right'
    | 'middle-left'
    | 'middle-center'
    | 'middle-right'
    | 'bottom-left'
    | 'bottom-center'
    | 'bottom-right';
}
interface ContentBlocksListItems {
  values?: PageContentValues[] | null;
}
// ...
export interface BackgroundColorConfig {
  values: BackgroundConfigItem[];
}

export interface BackgroundConfigItem {
  elements: {
    backgroundConfigItemElement: {
      values: BackgroundConfigItemElement[];
    };
  };
  type: 'Background Config';
}